+++ src/core/color.cpp
@@ -7,7 +7,7 @@ Color::Color()
   b=0;
 }
 
-Color::Color(double r, double g, double b){
+Color::Color(float r, float g, float b){
   this->r = r;
   this->g = g;
   this->b = b;
@@ -27,7 +27,7 @@ Color Color::operator*(Color& other)
 {
     return Color(other.r*r, other.g*g, other.b*b);
 }
-Color& Color::operator*=(double lambda)
+Color& Color::operator*=(float lambda)
 {
   r *= lambda;
   g *= lambda;
@@ -35,7 +35,7 @@ Color& Color::operator*=(double lambda)
   return *this;
 }
 
-Color Color::operator*(double lambda) const
+Color Color::operator*(float lambda) const
 {
   return Color(r*lambda, g*lambda, b*lambda);
 }
